:recycle: migrate to mariadb

+++ src/services/translator-activity-counts.ts
@@ -1,10 +1,10 @@
-import { type SQL, sql } from 'drizzle-orm';
+import { sql, type SQL } from 'drizzle-orm';
 
 import * as table from '../db/schema.js';
 
 export function translatorTradCountExpr(): SQL<number> {
   return sql<number>`coalesce((
-    select count(*)::int
+    select count(*)
     from ${table.gameTranslation}
     where ${table.gameTranslation.translatorId} = ${table.translator.id}
   ), 0)`;
@@ -12,7 +12,7 @@ export function translatorTradCountExpr(): SQL<number> {
 
 export function translatorReadCountExpr(): SQL<number> {
   return sql<number>`coalesce((
-    select count(*)::int
+    select count(*)
     from ${table.gameTranslation}
     where ${table.gameTranslation.proofreaderId} = ${table.translator.id}
   ), 0)`;
